fix: align DeFi token switch support OK-55458
